feat(driver): error processing, simple test upd
Closes RNDSTROPPY-55
